% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_converters.R
\name{allowed_class_to_sparta}
\alias{allowed_class_to_sparta}
\title{Classes that can be converted to sparta}
\usage{
allowed_class_to_sparta()
}
\description{
A non-argument function, that outputs the classes that can be converted to sparta
}
