% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_converters.R
\name{as_array}
\alias{as_array}
\alias{as_array.sparta}
\title{As array}
\usage{
as_array(x)

\method{as_array}{sparta}(x)
}
\arguments{
\item{x}{sparta object}
}
\value{
An array
}
\description{
Turn a sparse table into an array
}
\seealso{
\code{\link{as_array}}
}
