% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_converters.R
\name{as_df}
\alias{as_df}
\alias{as_df.sparta}
\title{As data frame}
\usage{
as_df(x, dense = FALSE)

\method{as_df}{sparta}(x, dense = FALSE)
}
\arguments{
\item{x}{sparta object}

\item{dense}{Logical indicating if zero cells should be present or not}
}
\value{
A data frame
}
\description{
Turn a sparse table into a data frame
}
\examples{
x <- array(
  c(1,0,0,2,3,4,0,0),
  dim = c(2,2,2),
  dimnames = list(
    a = c("a1", "a2"),
    b = c("b1", "b2"),
    c = c("c1", "c2")
  )
)

as_df(as_sparta(x))
}
\seealso{
\code{\link{as_array}}
}
