% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_gRNA_identity.R
\name{cell_gRNA_identity}
\alias{cell_gRNA_identity}
\title{Process sgRNA Data}
\usage{
cell_gRNA_identity(SUM159, min_count = 200, min_ratio = 0.7)
}
\arguments{
\item{SUM159}{A Seurat object containing GDO and motif data.}

\item{min_count}{Minimum count threshold for filtering (default is 200).}

\item{min_ratio}{Minimum ratio threshold for filtering (default is 0.7).}

\item{assay}{used to determine the identity of gRNAs in cells.}
}
\value{
A filtered Seurat object.
}
\description{
This function processes sgRNA data in a Seurat object, filtering based on count and ratio thresholds.
}
