% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality control.R
\name{cell_quality}
\alias{cell_quality}
\title{Quality control}
\usage{
cell_quality(
  SUM159,
  count_threshold = 100,
  feature_threshold = 200,
  percent_mt_threshold = 10,
  assay = "sgRNA",
  save_path = NULL
)
}
\arguments{
\item{SUM159}{A Seurat object containing GDO and motif data.}

\item{assay}{used to determine the identity of gRNAs in cells.}
}
\description{
This function for quality control of cells and to determine the thresholds for sgRNA identities in a Seurat object.
}
