% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dataDiagnostics}
\alias{dataDiagnostics}
\title{### Diagnostics functions ###
rm(list = ls())
# Create data
n <- 2000 #size of dataset
nyr <- 20 # number of years in data
nSamples <- 200 # set number of dates
useDates <- TRUE}
\usage{
dataDiagnostics(taxa, site, time_period)
}
\description{
# Create somes dates
first <- as.POSIXct(strptime("2003/01/01", "%Y/%m/%d"))
last <- as.POSIXct(strptime(paste(2003+(nyr-1),"/12/31", sep=''), "%Y/%m/%d"))
dt <- last-first
rDates <- first + (runif(nSamples)*dt)
}
\details{
# taxa are set as random letters
taxa <- sample(letters, size = n, TRUE)

# three sites are visited randomly
site <- sample(c('one', 'two', 'three'), size = n, TRUE)

# the date of visit is selected at random from those created earlier
if(useDates){
  time_period <- sample(rDates, size = n, TRUE)
} else {
  time_period <- sample(1:nSamples, size = n, TRUE)
}

dataDiagnostics(taxa, site, time_period)
}

