% Generated by roxygen2 (4.0.1): do not edit by hand
\name{date2timeperiod}
\alias{date2timeperiod}
\title{Assign dates to time periods}
\usage{
date2timeperiod(startDate, endDate, time_periods)
}
\arguments{
\item{startDate}{A vector of dates indicating the start date of a record}

\item{endDate}{A vector of dates indicating the end date of a record}

\item{time_period}{A data.frame with two columns, the first column gives the start year of each
time period and the second column gives the end year. These year ranges are inclusive.}
}
\value{
A vector, the same length as startDate and endDate giving the time period of each pair
of dates. Time period 1 is the time period with the earliest start year, 2 with the second
earliest start year and so on. If the dates don't fit in a time period or overlap a time period
they are given an NA value
}
\description{
This function assigns dates to timeperiods a nessecary step for methods that use time periods
rather than dates, such as Frescalo and Telfer
}
\examples{
# Load example dataset
data(ex_dat)

# Create time periods dataframe
time_periods <- data.frame(start=c(1980,1990),end=c(1989,1999))

# Create time periods column
ex_dat$timeperiod <- date2timeperiod(startDate = ex_dat$TO_STARTDATE,
                                     endDate = ex_dat$Date,
                                     time_periods = time_periods)
}

