% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_phenology.R
\name{detection_phenology}
\alias{detection_phenology}
\title{Diagnostics for the detectability with respect to Julian Date}
\usage{
detection_phenology(model, spname = NULL, bins = 12,
  density_function = TRUE)
}
\arguments{
\item{model}{a fitted sparta model of class \code{OccDet}.}

\item{spname}{optional name of the species (used for plotting)}

\item{bins}{number of points to estimate across the year. Defaults to 12}

\item{density_function}{whether the model used a density function to fit Julian date. This form was implemented from version 0.1.48 onwards. For models ran using earlier versions of the package this should be set to FALSE}
}
\value{
Some numbers.
}
\description{
Creates a plot of detectability over the season and calculates some simple statistics
}
\details{
Takes a object of \code{OccDet} fitted with the \code{jul_date} option

Calculate the phenology of detection and produces a plot of detectability over time for the reference data type.
}
\references{
van Strien, A.J., Termaat, T., Groenendijk, D., Mensing, V. & Kéry, M. (2010) 
            Site-occupancy models may offer new opportunities for dragonfly monitoring based on daily species lists. 
            \emph{Basic and Applied Ecology}, 11, 495–503.
}
