% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparta.R
\name{dim_names}
\alias{dim_names}
\alias{dim_names.sparta}
\title{Sparta dimension names}
\usage{
dim_names(x)

\method{dim_names}{sparta}(x)
}
\arguments{
\item{x}{sparta object}
}
\description{
Getter method for sparta dimension names
}
