% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObsModel.R
\name{getObsModel}
\alias{getObsModel}
\title{Create the observation model component of a sparta JAGS model}
\usage{
getObsModel(modeltype, verbose = FALSE)
}
\arguments{
\item{modeltype}{Character, one of: jul_date, catlistlength, contlistlength.
See \code{occDetFunc} for more information.}

\item{verbose}{Logical, if true progress is reported to the console}
}
\value{
A character, of JAGS model code, that describes the observation model.
}
\description{
This function is primarily for internal use within \code{getModelFile}. It is used to 
write an observation model that fits the users needs. The model is returned as a character.
}
