% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters.R
\name{getParameters}
\alias{getParameters}
\title{Modify the initial values object depending on the type of model we are running}
\usage{
getParameters(parameters, modeltype, verbose = FALSE)
}
\arguments{
\item{parameters}{A character vector of parameters you want to monitor.}

\item{modeltype}{Character, one of: indran, jul_date, catlistlength, contlistlength.
See \code{occDetFunc} for more information.}

\item{verbose}{Logical, if true progress is reported to the console}
}
\value{
A character, of JAGS model code, that describes the observation model.
}
\description{
This function is primarily for internal use within \code{occDetFunc}. It is used to 
return a vector of parameters to be monitored.
}
