% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparta.R
\name{vals}
\alias{vals}
\alias{vals.sparta}
\alias{get_values}
\alias{get_values.sparta}
\alias{dim_names}
\alias{dim_names.sparta}
\alias{names.sparta}
\title{Sparta getters}
\usage{
vals(x)

\method{vals}{sparta}(x)

get_values(x)

\method{get_values}{sparta}(x)

dim_names(x)

\method{dim_names}{sparta}(x)

\method{names}{sparta}(x)
}
\arguments{
\item{x}{sparta object}
}
\description{
Getter methods for sparta objects
}
