% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref_regions-data.R
\docType{data}
\name{gridref_regions}
\alias{gridref_regions}
\title{A dataset of UK grid references with regions}
\usage{
gridref_regions
}
\description{
This is a dataset of 1km-square UK grid references with regions ENGLAND, WALES, SCOTLAND, NORTHERN_IRELAND
}
\author{
Rob Cooke, 2021-02-15
}
