% Generated by roxygen2 (4.0.2): do not edit by hand
\name{occDetModel}
\alias{occDetModel}
\title{Occupancy detection models}
\usage{
occDetModel(taxa, site, time_period, print_progress = FALSE,
  species_list = unique(taxa), output_dir = getwd(), nyr = 3,
  n_iterations = 5000, burnin = 1500, thinning = 3, n_chains = 3,
  model.file = occDetBUGScode)
}
\arguments{
\item{taxa}{A character vector of taxon names, as long as the number of observations.}

\item{site}{A character vector of site names, as long as the number of observations.}

\item{time_period}{A numeric vector of user defined time periods, or a date vector,
as long as the number of observations.}

\item{species_list}{A list of taxa names for which models should be run. This is
optional and by default models will be run for all taxa}

\item{output_dir}{The output directory were the output for each taxa will be saved
as .rdata files. This will defualt to the working directory}

\item{nyr}{The minimum number of years on which a site must have records for it
to be included in the models}

\item{n_iterations}{An MCMC parameter - The number of interations}

\item{burnin}{An MCMC parameter - The length of the burn in}

\item{thinning}{An MCMC parameter - The thinning factor}

\item{n_chains}{An MCMC parameter - The number of chains to be run}

\item{model.file}{optionally a user defined BUGS model coded as a function (see ?jags,
       including the example there, for how this is done)}
}
\value{
A list of filepaths, one for each species run, giving the location of the
        output saved as a .rdata file, containing an object called 'out'
}
\description{
Run occupancy detection models as described in Isaac et al, 2014
}
\examples{
\dontrun{

# Create data
n <- 15000 #size of dataset
nyr <- 20 # number of years in data
nSamples <- 100 # set number of dates
nSites <- 50 # set number of sites

# Create somes dates
first <- as.POSIXct(strptime("2010/01/01", "\%Y/\%m/\%d"))
last <- as.POSIXct(strptime(paste(2010+(nyr-1),"/12/31", sep=''), "\%Y/\%m/\%d"))
dt <- last-first
rDates <- first + (runif(nSamples)*dt)

# taxa are set as random letters
taxa <- sample(letters, size = n, TRUE)

# three sites are visited randomly
site <- sample(paste('A', 1:nSites, sep=''), size = n, TRUE)

# the date of visit is selected at random from those created earlier
time_period <- sample(rDates, size = n, TRUE)

# run the model with these data for one species
results <- occDetModel(taxa = taxa,
                       site = site,
                       time_period = time_period,
                       species_list = c('a','m','g'),
                       n_iterations = 1000,
                       burnin = 10,
                       thinning = 2,
                       output_dir = "W:/PYWELL_SHARED/Pywell Projects/BRC/Tom August/R Packages/Trend analyses/occ_test_out")
}
}
\references{
Isaac, N.J.B., van Strien, A.J., August, T.A., de Zeeuw, M.P. and Roy, D.B. (2014).
            Statistics for citizen science: extracting signals of change from noisy ecological data.
            Methods in Ecology and Evolution, 5 (10), 1052-1060.

Roy, H.E., Adriaens, T., Isaac, N.J.B. et al. (2012) Invasive alien predator
            causes rapid declines of native European ladybirds. Diversity & Distributions,
            18, 717-725.
}
\keyword{bayesian,}
\keyword{distribution,}
\keyword{modeling}
\keyword{occupancy,}
\keyword{species,}
\keyword{trends,}

