% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrenceChange.R
\name{occurrenceChange}
\alias{occurrenceChange}
\title{Calculate percentage change between two years using Bayesian output}
\usage{
occurrenceChange(firstYear, lastYear, bayesOut)
}
\arguments{
\item{firstYear}{numeric, the first year over which the change is to be estimated}

\item{lastYear}{numeric, the last year over which the change is to be estimated}

\item{bayesOut}{occDet object as returned from occDetModel}
}
\value{
A list giving the mean, credible intervals and raw data from the
estimations.
}
\description{
Using the data returned from occDetModel this function models a linear
trend between two years for each iteration of the models. The predicted
values for the two years are then used to calculates a percentage
change. The results is a percentage change estimate for each of the
interations of the model. This distribution of the results is used to
calculate the mean estimate and the 95% credibale intervals.
}
\examples{
\dontrun{

#' # Create data
n <- 15000 #size of dataset
nyr <- 20 # number of years in data
nSamples <- 100 # set number of dates
nSites <- 50 # set number of sites

# Create somes dates
first <- as.Date(strptime("1980/01/01", "\%Y/\%m/\%d")) 
last <- as.Date(strptime(paste(1980+(nyr-1),"/12/31", sep=''), "\%Y/\%m/\%d")) 
dt <- last-first 
rDates <- first + (runif(nSamples)*dt)

# taxa are set as random letters
taxa <- sample(letters, size = n, TRUE)

# three sites are visited randomly
site <- sample(paste('A', 1:nSites, sep=''), size = n, TRUE)

# the date of visit is selected at random from those created earlier
time_period <- sample(rDates, size = n, TRUE)

# run the model with these data for one species
results <- occDetModel(taxa = taxa,
                       site = site,
                       time_period = time_period,
                       species_list = c('a','m','g'),
                       write_results = FALSE,
                       n_iterations = 1000,
                       burnin = 10,
                       thinning = 2)

 # estimate the change for one species      
 change <- occurrenceChange(firstYear = 1990,
                            lastYear = 1999,
                            bayesOut = results$a)   
}                   
}
