% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_apt_pro.R
\name{predict_apt_pro}
\alias{predict_apt_pro}
\title{Perform Motif Analysis and Prediction}
\usage{
predict_apt_pro(SUM159, assay = "motif1w", top_n = 20, save_path = NULL)
}
\arguments{
\item{SUM159}{A Seurat object containing GDO and motif data.}

\item{top_n}{Integer, number of top motifs to analyze.}

\item{save_path}{Plot save path. If NULL, plots are not saved.}

\item{A}{assay of the motif data.}
}
\value{
A data frame with predicted protein names and thresholds.
}
\description{
This function performs motif analysis, calculates differential motifs, filters them,
and determines thresholds using Gaussian mixture models. Optionally, saves plots.
}
