% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparta.R
\name{print.sparta}
\alias{print.sparta}
\title{Print}
\usage{
\method{print}{sparta}(x, ...)
}
\arguments{
\item{x}{sparta object}

\item{...}{For S3 compatability. Not used.}
}
\description{
Print method for sparta objects
}
