% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteSelectionMinL.r
\name{siteSelectionMinL}
\alias{siteSelectionMinL}
\title{List-length site selection}
\usage{
siteSelectionMinL(taxa, site, time_period, minL)
}
\arguments{
\item{taxa}{A character vector of taxon names}

\item{site}{A character vector of site names}

\item{time_period}{Anumeric vector of user defined time periods, or a date vector}

\item{minL}{numeric, The minimum number of taxa recorded at a site at a given time period 
(list-length) for the visit to be considered well sampled.}
}
\value{
A data.frame of data that forefills the selection criteria. This data has two attributes:
\code{visits} gives the total number of visits in the dataset (unique combinations of \code{site}
 and \code{time_period}), \code{success} gives the number of visits that satify the selection criteria
}
\description{
This function uses part of the method outlined in Roy et al (2012) and Isaac et al (2014) for selecting
well-sampled sites from a dataset using list length only. \code{\link{siteSelection}} is a wrapper
for this function that performs the complete site selection process as outlined in these papers.
}
\references{
needed
}
