% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/siteSelectionMinTP.r
\name{siteSelectionMinTP}
\alias{siteSelectionMinTP}
\title{Minimum time-period site selection}
\usage{
siteSelectionMinTP(taxa, site, time_period, minTP)
}
\arguments{
\item{taxa}{A character vector of taxon names}

\item{site}{A character vector of site names}

\item{time_period}{A numeric vector of user defined time periods, or a date vector}

\item{minTP}{numeric, The minimum number of time periods, or if time_period is a date the minimum
number of years, a site must be sampled in for it be be considered well sampled.}
}
\value{
A data.frame of data that forefills the selection criteria. This data has two attributes:
\code{sites} gives the total number of sites in the dataset, and \code{sucess} gives the number
of sites that satify the selection criteria
}
\description{
This function uses part of the method outlined in Roy et al (2012) and Isaac et al (2014) for selecting
well-sampled sites from a dataset using the number of time periods only. \code{\link{siteSelection}} is a wrapper
for this function that performs the complete site selection process as outlined in these papers.
}
\references{
needed
}

