% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sparta.R
\name{slice}
\alias{slice}
\alias{slice.sparta}
\title{Slice}
\usage{
slice(x, s)

\method{slice}{sparta}(x, s)
}
\arguments{
\item{x}{sparta object}

\item{s}{a slice in form of a named character vector}
}
\value{
A sparta object
}
\description{
Find the slice of a sparse table
}
\examples{

x <- array(
 c(1,0,0,2,3,4,0,0),
 dim = c(2,2,2),
 dimnames = list(
   a = c("a1", "a2"),
   b = c("b1", "b2"),
   c = c("c1", "c2")
 )
)

sx <- as_sparta(x)

# conditional probability table p(b,c|a)
sx <- as_cpt(sx, "a")

# the probability distriubtion when 'a' is 'a2' 
sxa2 <- slice(sx, c(a = "a2"))
get_val(sxa2, c(a = "a2", b = "b1", c = "c2"))

}
