% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparr.R
\docType{package}
\name{sparr-package}
\alias{sparr}
\alias{sparr-package}
\title{sparr: Representation and operations on sparse arrays}
\description{
To come...
}
\author{
\strong{Maintainer}: Mads Lindskou \email{ml@math.aau.dk}

Other contributors:
\itemize{
  \item Søren Højsgaard \email{sh@math.aau.dk} [contributor]
  \item Mikkel Meyer Andersen \email{ma@math.aau.dk} [contributor]
}

}
