% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparta.R
\name{sparsity}
\alias{sparsity}
\alias{sparsity.sparta}
\alias{sparsity.sparta_unity}
\title{Sparsity}
\usage{
sparsity(x)

\method{sparsity}{sparta}(x)

\method{sparsity}{sparta_unity}(x)
}
\arguments{
\item{x}{sparta}
}
\value{
The ratio of \code{ncol(x)} and the total statespace of \code{x}
}
\description{
Sparsity
}
\examples{

x <- array(
  c(1,0,0,2,3,4,0,0),
  dim = c(2,2,2),
  dimnames = list(
    a = c("a1", "a2"),
    b = c("b1", "b2"),
    c = c("c1", "c2")
  )
)

sx <- as_sparta(x)
sparsity(sx)
}
