\name{sparta}
\alias{sparta}
\title{The Sparta package}
\description{
  The Sparta package includes methods used to analyse
  trends in unstructured occurrence datasets and functions
  for mapping such data in the UK. Methods included in the 
  pckage include Frescalo, Telfer's change index, List-length
  models mixed models and others.\cr \cr
  Click on the index link below for more information on each 
  function  
}
