% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparta.R
\name{sparta_ones}
\alias{sparta_ones}
\title{Sparta Ones}
\usage{
sparta_ones(dim_names)
}
\arguments{
\item{dim_names}{A named list of discrete levels}
}
\value{
A sparta object
}
\description{
Construct a sparta object filled with ones
}
\examples{
sparta_ones(list(a = c("a1", "a2"), b = c("b1", "b2")))
}
