% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparta.R
\name{table_size}
\alias{table_size}
\alias{table_size.sparta}
\title{Number of elements in a table}
\usage{
table_size(x)

\method{table_size}{sparta}(x)
}
\arguments{
\item{x}{sparta}
}
\value{
The size of the sparta table \code{x}
}
\description{
Number of elements in a table
}
\examples{

x <- array(
  c(1,0,0,2,3,4,0,0),
  dim = c(2,2,2),
  dimnames = list(
    a = c("a1", "a2"),
    b = c("b1", "b2"),
    c = c("c1", "c2")
  )
)

sx <- as_sparta(x)
table_size(sx)
}
