% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/unicorns-data.r
\docType{data}
\name{unicorns}
\alias{unicorns}
\title{A fictional dataset of unicorn sightings}
\usage{
unicorns
}
\description{
This is a fictional occurrence dataset of 70 species of unicorn in the UK.
}
\author{
Tom August, 2015-07-01
}

