% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparta.R
\name{sum.sparta}
\alias{sum.sparta}
\alias{max.sparta}
\alias{min.sparta}
\alias{which_min_cell}
\alias{which_min_cell.sparta}
\alias{which_min_idx}
\alias{which_min_idx.sparta}
\alias{which_max_cell}
\alias{which_max_cell.sparta}
\alias{which_max_idx}
\alias{which_max_idx.sparta}
\title{Vector-like operations on sparta objects}
\usage{
\method{sum}{sparta}(x, ...)

\method{max}{sparta}(x, ...)

\method{min}{sparta}(x, ...)

which_min_cell(x)

\method{which_min_cell}{sparta}(x)

which_min_idx(x)

\method{which_min_idx}{sparta}(x)

which_max_cell(x)

\method{which_max_cell}{sparta}(x)

which_max_idx(x)

\method{which_max_idx}{sparta}(x)
}
\arguments{
\item{x}{sparta}

\item{...}{For S3 compatability.}
}
\description{
Vector-like operations on sparta objects
}
