% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_aptamer_difference.R
\name{visualize_aptamer_difference}
\alias{visualize_aptamer_difference}
\title{Visualize Prediction Results}
\usage{
visualize_aptamer_difference(predict_result, clust, save_path = NULL)
}
\arguments{
\item{clust}{A character string specifying the cluster to use.}

\item{save_path}{A character string specifying the path to save the plot. If NULL, the plot will not be saved.}

\item{df}{A data frame containing the prediction results.}
}
\value{
A ggplot object.
}
\description{
This function generates a plot for prediction results and saves it if a path is provided.
}
