% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDiagnostics.r
\name{dataDiagnostics}
\alias{dataDiagnostics}
\title{Data Diagnostics}
\usage{
dataDiagnostics(taxa, site, time_period, silent = FALSE)
}
\arguments{
\item{taxa}{A character vector of taxon names, as long as the number of observations.}

\item{site}{A character vector of site names, as long as the number of observations.}

\item{time_period}{A numeric vector of user defined time periods, or a date vector,
as long as the number of observations.}

\item{silent}{Logical, if \code{TRUE} plots will not be rendered and model results
       will not be printed to the console}
}
\value{
A list of filepaths, one for each species run, giving the location of the
        output saved as a .rdata file, containing an object called 'out'
}
\description{
This function provides visualisations of how the number of records in the
dataset changes over time and how the number of species recorded on a
visit changes over time. For each of these an linear model is run to test
if there is a significant trend.
}
\examples{
\dontrun{

### Diagnostics functions ###
rm(list = ls())
# Create data
n <- 2000 #size of dataset
nyr <- 20 # number of years in data
nSamples <- 200 # set number of dates
useDates <- TRUE

# Create somes dates
first <- as.POSIXct(strptime("2003/01/01", "\%Y/\%m/\%d"))
last <- as.POSIXct(strptime(paste(2003+(nyr-1),"/12/31", sep=''), "\%Y/\%m/\%d"))
dt <- last-first
rDates <- first + (runif(nSamples)*dt)

# taxa are set as random letters
taxa <- sample(letters, size = n, TRUE)

# three sites are visited randomly
site <- sample(c('one', 'two', 'three'), size = n, TRUE)

# the date of visit is selected at random from those created earlier
if(useDates){
  time_period <- sample(rDates, size = n, TRUE)
} else {
  time_period <- sample(1:nSamples, size = n, TRUE)
}

dataDiagnostics(taxa, site, time_period)
}
}

