% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMetrics.R
\name{dataMetrics}
\alias{dataMetrics}
\title{Calculate dataset metrics}
\usage{
dataMetrics(sp, formattedData, suffix = NULL)
}
\arguments{
\item{sp}{character string identifying the species name}

\item{formattedData}{output from formattedOccData}

\item{suffix}{Not used}
}
\description{
Takes a formattedOccData object and the name of a species. Runs a set of calculations on the dataset for that species
Based on rules of thumb. Original version by Mark Logie; integrated into sparta by NJBI
}
\references{
Pocock, Logie, Isaac, Outhwaite & August. Rapid assessment of the suitability of multi-species citizen science datasets for occupancy trend analysis. bioRxiv 813626 (2019) doi:10.1101/813626.
}
