% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/frescalo.R
\name{frescalo}
\alias{frescalo}
\title{Frescalo trend analysis}
\usage{
frescalo(Data, frespath, time_periods, site_col, sp_col, year_col = NULL,
  start_col = NULL, end_col = NULL, species_to_include = NULL,
  sinkdir = NULL, plot_fres = FALSE, Fres_weights = "LCGB",
  non_benchmark_sp = NULL, fres_site_filter = NULL, phi = 0.74,
  alpha = 0.27, trend_option = "arithmetic", NYears = 10,
  ignore.ireland = F, ignore.channelislands = F)
}
\arguments{
\item{Data}{A dataframe object. This should consist of rows of
observations and columns indicating the species and location as well as
either the year of the observation or columns specifying the start and end
dates of the observation. It is important that date columns are in
a date format.}

\item{frespath}{the path to the frescalo .exe file. This can be downloaded from
http://www.brc.ac.uk/biblio/frescalo-computer-program-analyse-your-biological-records.
It is currently only available for Windows. The directory where the .exe
is saved should be writeable.}

\item{time_periods}{A dataframe object with two columns. The first column contains the
start year of each time period and the second column contains the end year of
each time period. Time periods should not overlap.}

\item{site_col}{The name of the site column in \code{Data}}

\item{sp_col}{The name of the species column in \code{Data}}

\item{year_col}{The name of the year column in \code{Data}}

\item{start_col}{The name of the start date column in \code{Data}}

\item{end_col}{The name of the end date column in \code{Data}}

\item{species_to_include}{Optionally a character vector listing the names of species to be used.
Species not in your list are ignored. This is useful if you are only interested in a
subset of species.}

\item{sinkdir}{String giving the output directory for results}

\item{plot_fres}{Logical, if \code{TRUE} maps are produced by Frescalo. Default is
\code{FALSE}. CURRENTLY ONLY WORKS FOR UK GRID-REFERENCE DATA}

\item{Fres_weights}{'LC*' specifies a weights files based on landcover data.
The suffix specifies the extend ('LCUK', 'LCNI' or 'LCGB'). 'VP' uses a weights
file based on vascular plant data for the UK
, both are included in the package. Alternativly a custom weights file can be
given as a data.frame. This must have three columns: target cell, neighbour cell,
weight. Default is 'LCGB'}

\item{non_benchmark_sp}{a character vector, giving the names of species not to be
used as benchmarks in Frescalo. Default is \code{NULL} and all species are used.
See Hill, 2011 for reasons why some species may not be suitable benchmarks.}

\item{fres_site_filter}{Optionally a character vector giving
the names of sites to be used in the trend analysis. Sites not include in this
list are not used for estimating TFactors. Default is \code{NULL} and all sites are
used.}

\item{phi}{Target frequency of frequency-weighted mean frequency. Default is 0.74 as in
Hill (2011). If this value is smaller than the 98.5 percentile of input phi it is
automatically increased and a warning message is generated. This is limited to
0.50 to 0.95.}

\item{alpha}{the proportion of the expected number of species in a cell to be treated as
benchmarks. Default is 0.27 as in Hill (2011). This is limited to 0.08 to 0.50.}

\item{trend_option}{Set the method by which you wish to calculate percentage change. This can currently
be set to either \code{'arithmetic'} (default) or \code{'geometric'}. Arimthmetic calculates
percentage change in a linear fashion such that a decline of 50\% over 50 years is
equal to 10\% in 10 years. Using the same example a Geometric trend would be 8.44\%
every 10 years as this work on a compound rate.}

\item{NYears}{The number of years over which you want the percentage change to be calculated (i.e.
10 gives a decadal change). Default = 10}

\item{ignore.ireland}{Logical, if \code{TRUE} Irish hectads are removed. Default
is \code{FALSE}}

\item{ignore.channelislands}{Logical, if \code{TRUE} channel island hectads are
removed. Default is \code{FALSE}}
}
\value{
Results are saved to file and most are returned in a list to R.

        The list object returned is comprised of the following:

        \item{\bold{$paths}}{This list of file paths provides the locations of the raw data files
        for $log, $stat, $freq and $trend, in that order}

        \item{\bold{$trend}}{This dataframe provides the list of time factors for each species}

         \tabular{rll}{
         - \tab \code{Species} \tab Name of species\cr
         - \tab \code{Time} \tab Time period, specified as a class (e.g. 1970); times need not be numeric and are indexed as character strings\cr
         - \tab \code{TFactor} \tab Time factor, the estimated relative frequency of species at the time\cr
         - \tab \code{St_Dev} \tab Standard deviation of the time factor, given that spt (defined below) is a weighted sum of binomial variates\cr
         - \tab \code{X} \tab Number of occurrences of species at the time period\cr
         - \tab \code{Xspt} \tab Number of occurrences, given reduced weight of locations having very low sampling effort\cr
         - \tab \code{Xest} \tab Estimated number of occurrences; this should be equal to spt if the algorithm has converged\cr
         - \tab \code{N>0.00} \tab Number of locations with non-zero probability of the species occurring\cr
         - \tab \code{N>0.98} \tab Number of locations for which the probability of occurrence was estimated as greater than 0.98\cr
         }

       \item{\bold{$stat}}{Location report}

         \tabular{rll}{
         - \tab \code{Location} \tab Name of location; in this case locations are hectads of the GB National Grid \cr
         - \tab \code{Loc_no} \tab Numbering (added) of locations in alphanumeric order \cr
         - \tab \code{No_spp} \tab Number of species at that location; the actual number which may be zero \cr
         - \tab \code{Phi_in} \tab Initial value of phi, the frequency-weighted mean frequency \cr
         - \tab \code{Alpha} \tab Sampling effort multiplier (to achieve standard value of phi) \cr
         - \tab \code{Wgt_n2} \tab effective number N2 for the neighbourhood weights; this is small if there are few floristically similar hectads close to the target hectad.  It is (sum weights)^2 / (sum weights^2) \cr
         - \tab \code{Phi_out} \tab Value of phi after rescaling; constant, if the algorithm has converged\cr
         - \tab \code{Spnum_in} \tab Sum of neighbourhood frequencies before rescaling\cr
         - \tab \code{Spnum_out} \tab Estimated species richness, i.e. sum of neighbourhood frequencies after rescaling\cr
         - \tab \code{Iter} \tab Number of iterations for algorithm to converge\cr
         }

        \item{\bold{$freq}}{Listing of rescaled species frequencies}

         \tabular{rll}{
         - \tab \code{Location} \tab Name of location\cr
         - \tab \code{Species} \tab Name of species\cr
         - \tab \code{Pres} \tab Record of species in location (1 = recorded, 0 = not recorded)\cr
         - \tab \code{Freq} \tab Frequency of species in neighbourhood of location\cr
         - \tab \code{Freq_1} \tab Estimated probabilty of occurrence, i.e. frequency of species after rescaling\cr
         - \tab \code{SD_Frq1} \tab Standard error of Freq_1, calculated on the assumption that Freq is a binomial variate with standard error sqrt(Freq*(1-Freq)/ Wgt_n2), where Wgt_n2 is as defined for samples.txt in section (b)\cr
         - \tab \code{Rank} \tab Rank of frequency in neighbourhood of location\cr
         - \tab \code{Rank_1} \tab Rescaled rank, defined as Rank/Estimated species richness\cr
         }

        \item{\bold{$log}}{This records all the output sent to the console when running frescalo}

        \item{\bold{$lm_stats}}{The results of linear modelling of TFactors}

         \tabular{rll}{
         - \tab \code{SPECIES} \tab Name of species used internally by frescalo\cr
         - \tab \code{NAME} \tab Name of species as appears in raw data\cr
         - \tab \code{b} \tab The slope of the model\cr
         - \tab \code{a} \tab The intercept\cr
         - \tab \code{b_std_err} \tab Standard error of the slope\cr
         - \tab \code{b_tval} \tab t-value for a test of significance of the slope\cr
         - \tab \code{b_pval} \tab p-value for a test of significance of the slope\cr
         - \tab \code{a_std_err} \tab Standard error of the intercept\cr
         - \tab \code{a_tval} \tab t-value for a test of significance of the intercept\cr
         - \tab \code{a_pval} \tab p-value for a test of significance of the intercept\cr
         - \tab \code{adj_r2} \tab Rescaled rank, defined as Rank/Estimated species richness\cr
         - \tab \code{r2} \tab t-value for a test of significance of the intercept\cr
         - \tab \code{F_val} \tab F-value of the model\cr
         - \tab \code{F_num_df} \tab Degrees of freedom of the model\cr
         - \tab \code{F_den_df} \tab Denominator degrees of freedom from the F-statistic\cr
         - \tab \code{Ymin} \tab The earliest year in the dataset\cr
         - \tab \code{Ymax} \tab The latest year in the dataset\cr
         - \tab \code{change_...} \tab The percentage change dependent on the values given to \code{trend_option} and \code{NYears}.\cr
         }
         \bold{The following columns are only produced when there are only two time periods}
         \tabular{rll}{
         - \tab \code{Z_VAL} \tab Z-value for the significance test of the trend\cr
         - \tab \code{SIG_95} \tab A logical statement indicating if the trend is significant (TRUE) or non-significant (FALSE)\cr
         }
}
\description{
A function for using Frescalo (Hill, 2011), a tool for analysing occurrence data when
recording effort is not known. This function returns the output from Frescalo to the
R session and saves it to the path specified by \code{sinkdir}. By setting
\code{plot_fres} to \code{TRUE} maps of the results will also be saved. Plotting the
returned object gives a useful summary.
}
\examples{
\dontrun{
# Load data
data(unicorns)

# Run frescalo (data is save to the working directory as sinkdir is not given)
fres_out <- frescalo(Data = unicorns,
                     time_periods = data.frame(start=c(1980,1990),end=c(1989,1999)),
                     site_col = 'hectad',
                     sp_col = 'CONCEPT',
                     start_col = 'TO_STARTDATE',
                     end_col = 'Date')
}
}
\references{
Hill, Mark. Local frequency as a key to interpreting species occurrence data when
recording effort is not known. 2011. \emph{Methods in Ecology and Evolution}, 3 (1), 195-205.
}
\keyword{frescalo}
\keyword{trends,}

