% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlSummary.R
\name{htmlSummary}
\alias{htmlSummary}
\title{Create HTML Report}
\usage{
htmlSummary(occDet, open = TRUE, output_dir = getwd(), output_file = NULL, ...)
}
\arguments{
\item{occDet}{An object of class occDet}

\item{open}{Logical, if \code{TRUE} the html report is opened once complete}

\item{output_dir}{Character, the directory where the html file will be saved,
defaults to the working directory.}

\item{output_file}{(Optional) The file name given to the html file}

\item{...}{Additional arguments passed to rmarkdown::render}
}
\value{
Path to html report
}
\description{
Create HTML Report for an \code{occDet} object.
}
