% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DetectionOverTime.R
\name{plot_DetectionOverTime}
\alias{plot_DetectionOverTime}
\title{Diagnostics for the detection model with respect to Length}
\usage{
plot_DetectionOverTime(model, spname = NULL, min.yr = NULL, CI = 95)
}
\arguments{
\item{model}{a fitted sparta model of class \code{OccDet}.}

\item{spname}{optional name of the species (used for plotting)}

\item{min.yr}{Minimum year for the detection (e.g., 1990).}

\item{CI}{Confidence interval}
}
\value{
This function returns plot showing the detection probability on the y axis and year on the x.
}
\description{
Creates a plot of detectability by year for differing list lengths from an occupancy model output.
}
\details{
Takes a object of \code{OccDet}

Calculates the detection probability and produces a plot of detectability over time for the reference data type.
}
