% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparta-package.R
\docType{package}
\name{sparta-package}
\alias{sparta}
\alias{sparta-package}
\title{sparta: Sparse Tables}
\description{
Fast Multiplication and Marginalization of Sparse Tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlindsk/sparta}
  \item Report bugs at \url{https://github.com/mlindsk/sparta/issues}
}

}
\author{
\strong{Maintainer}: Mads Lindskou \email{madslindskou@gmail.com}

}
