% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparta.r
\docType{package}
\name{sparta}
\alias{sparta-package}
\alias{sparta}
\title{\pkg{sparta} Trend Analysis for Unstructured Data}
\description{
The Sparta package includes methods used to analyse trends in
unstructured occurrence datasets. Methods included in the package
include Frescalo, Telfer's change index, Reporting rate models
and Bayesian Occupancy models. These methods are reviewed in 
Issac et al (2014), available at \url{http://onlinelibrary.wiley.com/doi/10.1111/2041-210X.12254/abstract}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BiologicalRecordsCentre/sparta}
  \item Report bugs at \url{https://github.com/BiologicalRecordsCentre/sparta/issues}
}

}
\author{
\strong{Maintainer}: Dylan Carbone \email{dylcar@ceh.ac.uk} (\href{https://orcid.org/0009-0003-5290-786X}{ORCID})

Authors:
\itemize{
  \item Nick Isaac (\href{https://orcid.org/0000-0002-4869-8052}{ORCID})
  \item Tom August \email{tomaug@ceh.ac.uk} (\href{https://orcid.org/0000-0003-1116-3385}{ORCID})
  \item Colin Harrower (\href{https://orcid.org/0000-0001-5070-5293}{ORCID})
  \item Jack Hatfield (\href{https://orcid.org/0000-0002-6361-0629}{ORCID})
  \item Mark Hill (\href{https://orcid.org/0000-0002-3323-4458}{ORCID})
  \item Francesca Mancini (\href{https://orcid.org/0000-0003-4085-4978}{ORCID})
  \item Charlie Outhwaite (\href{https://orcid.org/0000-0001-9997-6780}{ORCID})
  \item Gary Powney (\href{https://orcid.org/0000-0003-3313-7786}{ORCID})
}

}
\keyword{internal}
