% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/sparta.r
\docType{package}
\name{sparta}
\alias{sparta}
\alias{sparta-package}
\title{\pkg{sparta} Trend Analysis for Unstructured Data}
\description{
The Sparta package includes methods used to analyse trends in
unstructured occurrence datasets. Methods included in the package
include Frescalo, Telfer's change index, Reporting rate models
and Bayesian Occupancy models. These methods are reviewed in
Issac et al (2014), available at \url{http://onlinelibrary.wiley.com/doi/10.1111/2041-210X.12254/abstract}
}

