% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparta.R
\name{sparta_struct}
\alias{sparta_struct}
\title{Construct sparta object}
\usage{
sparta_struct(x, vals, dim_names)
}
\arguments{
\item{x}{matrix where columns represents cells in an array-like object}

\item{vals}{vector of values corresponding to x}

\item{dim_names}{a named list}
}
\value{
A sparta object
}
\description{
Helper function to construct a sparta object with given values and dim names
}
