% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/telfer.R
\name{telfer}
\alias{telfer}
\title{Telfer's change index}
\usage{
telfer(taxa, site, time_period, minSite = 5, useIterations = TRUE,
  iterations = 10)
}
\arguments{
\item{taxa}{A character vector of taxon names}

\item{site}{A character vector of site names}

\item{time_period}{A numeric vector of user defined time periods, or a date vector}

\item{minSite}{The minimum number of sites occupied in the first time period in
order for a trend to be calculated for a taxon.}

\item{useIterations}{A logical variable indicating whether iterations should be used.
Iterations are used to account for increased variation in the logit proportions
close to zero and one (see Telfer et al 2002). Default is \code{TRUE}}

\item{iterations}{If \code{useIterations} is \code{TRUE}, then this parameter indicates
the number of iterations to be used. In Telfer et al 2002 the number of iterations
used were 7 and 8 for the two datasets for which it was applied. The defualt here
is 10.}
}
\description{
Telfers change index is designed to assess the relative change in range size of species
between two time periods (Telfer et al 2002). This function can take multiple time periods
and will complete all pairwise comparisons.
}
\examples{
# Create fake data
SS <- 1000 # number of observations
taxa <- sample(letters, SS, replace = TRUE)
site <- sample(paste('A', 1:20, sep = ''), SS, replace = TRUE)
time_period <- sample(1:3, SS, replace = TRUE)

TelferResult <- telfer(taxa, site, time_period)
}
\references{
Telfer, M.G., Preston, C.D., & Rothery, P. (2002) A general method for
            measuring relative change in range size from biological atlas data.
            Biological Conservation, 107, 99-109.
}

