% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicorns-data.r
\docType{data}
\name{unicorns}
\alias{unicorns}
\title{A fictional dataset of unicorn sightings}
\usage{
unicorns
}
\description{
This is a fictional occurrence dataset of 20 species of unicorn in the UK. The dataset has column names, start_date, end_date, site and species respectively, corresponding to the start datetime of a occurence, end datetime of the occurence, site ID, and species ID.
}
\author{
Tom August, 2015-07-01
}
